[HaCreator] MapSimulator: parity ownership and runtime packet updates
Commit parity-focused MapSimulator code updates across character, field, interaction, manager, pool, and UI runtime paths, including new packet/session parity helpers.

+++ HaCreator/MapSimulator/Managers/PacketOwnedRewardResultRuntime.cs
@@ -94,13 +94,13 @@ namespace HaCreator.MapSimulator.Managers
             }
         }
 
-        public static string FormatMesoGiveSucceededText(int mesoAmount)
+        public static string FormatMesoGiveSucceededText(uint mesoAmount)
         {
             string format = ResolveTextFormat(
                 MesoGiveSucceededStringPoolId,
                 "You have received {0:N0} mesos.",
                 1);
-            return FormatInvariant(format, mesoAmount);
+            return FormatInvariant(format, (ulong)mesoAmount);
         }
 
         public static string GetMesoGiveFailedText()
